+++ scripts/ageing_society_cfps/cfps_fill_missing_data.py
@@ -43,9 +43,14 @@ logging.basicConfig(
 logging.info("Configure module")
 current_module = "ageing_society"
 
+logging.info("Configure data source")
+data_source = "cfps"
+
 logging.info("Configure paths")
-path_data_raw = data_home / "raw_data" / current_module / current_version
-path_data_clean = data_home / "clean_data" / current_module / current_version
+path_data_raw = data_home / "raw_data" / current_module / current_version / data_source
+path_data_clean = (
+    data_home / "clean_data" / current_module / current_version / data_source
+)
 
 if not path_data_clean.exists():
     path_data_clean.mkdir(parents=True, exist_ok=True)